requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
alembic==1.13.2
amqp==5.2.0
aniso8601==9.0.1
annotated-types==0.7.0
anyio==3.7.1
async-timeout==4.0.3
attrs==24.2.0
billiard==4.2.0
blinker==1.8.2
cachetools==5.5.0
celery==5.4.0
certifi==2024.8.30
charset-normalizer==2.0.12
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
distro==1.9.0
exceptiongroup==1.2.2
Flask==2.1.0
Flask-Cors==5.0.0
Flask-Mail==0.10.0
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
frozenlist==1.4.1
google-ai-generativelanguage==0.6.9
google-api-core==2.19.2
google-api-python-client==2.145.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.0
googleapis-common-protos==1.65.0
greenlet==3.1.0
grpcio==1.66.1
grpcio-status==1.66.1
gunicorn==20.1.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.2
idna==3.8
importlib_metadata==8.5.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
kombu==5.4.1
Mako==1.3.5
MarkupSafe==2.1.5
multidict==6.1.0
openai==1.44.1
packaging==24.1
pluggy==1.5.0
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==5.28.1
py==1.11.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.1
pydantic_core==2.23.3
PyJWT==2.9.0
pyparsing==3.1.4
pytest==7.1.2
python-dateutil==2.9.0.post0
python-dotenv==0.19.2
pytz==2024.2
redis==5.0.8
requests==2.27.1
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==1.4.46
tomli==2.0.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.20
vine==5.1.0
wcwidth==0.2.13
Werkzeug==2.0.3
yarl==1.11.1
zipp==3.20.1
